+++ CHANGES.md
@@ -5,7 +5,8 @@
 * 
 
 **Fixed:**
-* 
+* Project's not properly loading custom control classes' inherited properties
+* custom control classes weren't storing inherited properties when initialized from a specification
 
 **Notes:**
 * June 23: